LayoutTests/ietestcenter/Javascript/15.2.3.14-3-4-expected.txt

15.2.3.14-3-4 (TestCases/chapter15/15.2/15.2.3/15.2.3.14/15.2.3.14-3-4.js) Object.keys of an arguments object returns the indices of the given arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
